+++ src/controls/default_controls.ts
@@ -1,112 +1,6 @@
 import { FabricObject } from '../shapes/Object/FabricObject';
 import { Textbox } from '../shapes/Textbox';
-import { scaleCursorStyleHandler, scalingEqually } from './scale';
-import { changeWidth } from './changeWidth';
-import { rotationStyleHandler, rotationWithSnapping } from './rotate';
-import {
-  scaleOrSkewActionName,
-  scaleSkewCursorStyleHandler,
-  scalingXOrSkewingY,
-  scalingYOrSkewingX,
-} from './scaleSkew';
-import { Control } from './Control';
-
-// use this function if you want to generate new controls for every instance
-export const createObjectDefaultControls = () => ({
-  ml: new Control({
-    x: -0.5,
-    y: 0,
-    cursorStyleHandler: scaleSkewCursorStyleHandler,
-    actionHandler: scalingXOrSkewingY,
-    getActionName: scaleOrSkewActionName,
-  }),
-
-  mr: new Control({
-    x: 0.5,
-    y: 0,
-    cursorStyleHandler: scaleSkewCursorStyleHandler,
-    actionHandler: scalingXOrSkewingY,
-    getActionName: scaleOrSkewActionName,
-  }),
-
-  mb: new Control({
-    x: 0,
-    y: 0.5,
-    cursorStyleHandler: scaleSkewCursorStyleHandler,
-    actionHandler: scalingYOrSkewingX,
-    getActionName: scaleOrSkewActionName,
-  }),
-
-  mt: new Control({
-    x: 0,
-    y: -0.5,
-    cursorStyleHandler: scaleSkewCursorStyleHandler,
-    actionHandler: scalingYOrSkewingX,
-    getActionName: scaleOrSkewActionName,
-  }),
-
-  tl: new Control({
-    x: -0.5,
-    y: -0.5,
-    cursorStyleHandler: scaleCursorStyleHandler,
-    actionHandler: scalingEqually,
-  }),
-
-  tr: new Control({
-    x: 0.5,
-    y: -0.5,
-    cursorStyleHandler: scaleCursorStyleHandler,
-    actionHandler: scalingEqually,
-  }),
-
-  bl: new Control({
-    x: -0.5,
-    y: 0.5,
-    cursorStyleHandler: scaleCursorStyleHandler,
-    actionHandler: scalingEqually,
-  }),
-
-  br: new Control({
-    x: 0.5,
-    y: 0.5,
-    cursorStyleHandler: scaleCursorStyleHandler,
-    actionHandler: scalingEqually,
-  }),
-
-  mtr: new Control({
-    x: 0,
-    y: -0.5,
-    actionHandler: rotationWithSnapping,
-    cursorStyleHandler: rotationStyleHandler,
-    offsetY: -40,
-    withConnection: true,
-    actionName: 'rotate',
-  }),
-});
-
-export const createResizeControls = () => ({
-  mr: new Control({
-    x: 0.5,
-    y: 0,
-    actionHandler: changeWidth,
-    cursorStyleHandler: scaleSkewCursorStyleHandler,
-    actionName: 'resizing',
-  }),
-  ml: new Control({
-    x: -0.5,
-    y: 0,
-    actionHandler: changeWidth,
-    cursorStyleHandler: scaleSkewCursorStyleHandler,
-    actionName: 'resizing',
-  }),
-});
-
-export const createTextboxDefaultControls = () => ({
-  ...createObjectDefaultControls(),
-  ...createResizeControls(),
-});
-
-export const defaultControls = createObjectDefaultControls();
+import { createResizeControls, defaultControls } from './defaultControls';
 
 // shared with the default object on purpose
 export const textboxDefaultControls = {
